pyproject.toml
[project]
name = "ese-cli"
version = "1.0.0"
description = "Ensemble Software Engineering (ESE) CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "questionary>=2.0",
  "PyYAML>=6.0",
  "pydantic>=2.7",
]

[project.scripts]
ese = "ese.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]
